+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Fixed **format** to use git as a default value.
 * Updated the **upload** command to support reports.
 * Fixed an issue where **generate-docs** command was displaying 'None' when credentials parameter display field configured was not configured.
+* Added to the **validate** command validation for test playbook having default version.
 
 # 1.3.9
 * Added a validation verifying that the pack's README.md file is not equal to pack description.